divided utilities
