+++ python-codes/Outputs/simulation.log
@@ -1,102 +1,102 @@
 step Epot press
-0 109859.37 3746077.07
-10 109851.57 3746077.22
-20 109570.09 3746077.20
-30 109547.89 3746077.20
-40 2313.63 67694.46
-50 2299.96 67694.41
-60 2298.10 67694.43
-70 2132.42 67694.47
-80 590.13 14558.01
-90 588.83 14558.10
-100 498.53 14558.19
-110 458.58 14558.19
-120 455.12 14558.20
-130 61.76 935.27
-140 57.03 935.27
-150 56.32 935.13
-160 56.10 935.16
-170 53.40 935.01
-180 53.33 935.02
-190 50.25 850.65
-200 50.29 850.65
-210 23.01 690.33
-220 21.35 690.43
-230 11.51 690.33
-240 -0.32 281.43
-250 -2.79 281.85
-260 -2.41 289.92
-270 -2.52 289.97
-280 -2.98 290.03
-290 -3.12 290.03
-300 -3.00 289.96
-310 -3.33 290.00
-320 -3.38 289.95
-330 -1.41 289.46
-340 -1.00 289.46
-350 -1.17 289.45
-360 -1.02 289.49
-370 -2.09 255.33
-380 -2.01 255.31
-390 -2.07 255.33
-400 -3.46 257.83
-410 -3.36 257.84
-420 -3.61 257.78
-430 -3.67 257.77
-440 -3.92 257.76
-450 -3.95 257.76
-460 -3.91 261.39
-470 -4.74 249.63
-480 -4.81 249.55
-490 -4.79 250.16
-500 -4.10 249.56
-510 -4.14 249.56
-520 -4.18 247.31
-530 -4.13 248.23
-540 -4.13 248.22
-550 -5.08 248.21
-560 -4.94 248.24
-570 -4.75 247.11
-580 -4.61 247.31
-590 -3.55 247.32
-600 -3.00 247.30
-610 -2.81 247.31
-620 -3.37 247.46
-630 -3.93 247.45
-640 -4.07 247.45
-650 -3.83 247.45
-660 -3.68 247.46
-670 -3.72 247.46
-680 -3.70 247.50
-690 -3.46 247.41
-700 -2.88 247.41
-710 -3.64 247.60
-720 -3.55 247.55
-730 -3.23 247.26
-740 -3.56 247.60
-750 -4.15 247.85
-760 -3.64 248.05
-770 -3.34 248.12
-780 -3.46 248.14
-790 -3.38 248.14
-800 -2.45 248.11
-810 -2.17 248.10
-820 -2.14 248.25
-830 -1.80 248.26
-840 -2.31 248.48
-850 -2.40 248.48
-860 -2.58 248.28
-870 -2.53 248.32
-880 -2.73 248.32
-890 -2.83 248.20
-900 -4.02 248.23
-910 -3.56 248.14
-920 -3.83 248.36
-930 -3.93 248.33
-940 -3.92 248.31
-950 -3.91 248.49
-960 -4.17 248.66
-970 -3.95 248.64
-980 -3.74 248.69
-990 -4.64 248.61
-1000 -4.41 248.62
+0 8794.69 248.50
+10 7421.00 248.50
+20 7406.55 248.41
+30 7403.86 248.43
+40 7400.42 248.43
+50 7400.43 248.46
+60 7400.26 248.51
+70 7400.34 248.64
+80 7386.56 248.95
+90 7385.30 248.95
+100 7384.86 248.95
+110 913.20 248.95
+120 913.54 248.94
+130 882.35 248.96
+140 785.34 248.92
+150 783.15 249.77
+160 603.01 249.77
+170 293.46 249.71
+180 265.51 249.68
+190 116.08 249.67
+200 116.55 249.70
+210 116.17 249.68
+220 115.75 250.44
+230 115.64 250.43
+240 114.71 250.43
+250 111.74 248.30
+260 111.62 248.40
+270 112.00 248.44
+280 107.61 248.44
+290 107.48 248.47
+300 107.57 248.31
+310 91.24 249.18
+320 91.41 249.18
+330 91.34 249.22
+340 45.01 249.08
+350 43.72 249.07
+360 10.11 249.07
+370 9.89 248.61
+380 9.56 248.56
+390 9.65 248.55
+400 4.54 265.39
+410 0.13 265.17
+420 0.27 265.12
+430 0.01 251.22
+440 -2.12 251.15
+450 -4.15 251.17
+460 -4.54 251.21
+470 -4.44 251.21
+480 -4.21 251.25
+490 -3.76 251.25
+500 -4.23 251.12
+510 -3.96 249.15
+520 -3.68 261.26
+530 -2.76 261.21
+540 -3.35 261.20
+550 -3.46 261.26
+560 -3.54 261.26
+570 -3.53 261.39
+580 -3.42 261.24
+590 -3.25 261.23
+600 -4.40 255.97
+610 -4.51 252.83
+620 -4.39 252.84
+630 -4.07 252.90
+640 -4.07 249.18
+650 -4.33 249.32
+660 -3.98 249.31
+670 -2.75 249.32
+680 -2.97 249.32
+690 -3.19 249.34
+700 -3.31 248.42
+710 -3.39 248.41
+720 -3.37 247.50
+730 -3.96 247.42
+740 -4.48 247.45
+750 -4.37 247.48
+760 -4.39 247.48
+770 -4.48 247.45
+780 -4.43 247.49
+790 -4.51 247.61
+800 -4.38 247.61
+810 -4.15 247.58
+820 -3.73 247.66
+830 -3.82 247.77
+840 -3.74 248.08
+850 -3.95 248.12
+860 -3.81 247.98
+870 -2.74 247.91
+880 -2.75 248.01
+890 -2.78 248.04
+900 -1.87 248.03
+910 -2.31 248.03
+920 -2.21 248.05
+930 -1.41 248.02
+940 -1.45 248.02
+950 1.08 248.01
+960 0.50 248.01
+970 1.11 247.90
+980 0.58 247.96
+990 0.63 248.00
+1000 0.01 247.99
