+++ python-codes/Utilities.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from utilities import contact_matrix, compute_neighbor_lists
+from contact_utils import contact_matrix, compute_neighbor_lists
 
 
 class Utilities:
