+++ python-codes/test_8a.py
@@ -1,5 +1,3 @@
-
-
 from MonteCarlo import MonteCarlo
 from pint import UnitRegistry
 ureg = UnitRegistry()
