+++ python-codes/tools.py
@@ -1,5 +1,3 @@
-
-
 import numpy as np
 
 import logging
